+++ pcep/spi/src/main/java/org/opendaylight/protocol/pcep/spi/PSTUtil.java
@@ -21,7 +21,7 @@ public final class PSTUtil {
      * @return  true if setup is via RSVP-TE signaling protocol
      */
     public static boolean isDefaultPST(final PathSetupType pst) {
-        if (pst != null && pst.getPst() != null && pst.getPst().shortValue() != 0) {
+        if (pst != null && pst.getPst() != null && pst.getPst() != 0) {
             return false;
         }
         return true;
